// demo/merchant/src/models/types.ts
  export interface Address {
    /** Name of the person to whom the items are shipped */
    name: string;
    /** First line of address */
    line_one: string;
    /** Optional second line of address */
    line_two?: string;
    /** Address city/district/suburb/town/village */
    city: string;
    /** Address state/county/province/region */
    state?: string;
    /** Address country */
    country: string;
    /** Address postal code or zip code */
    postal_code: string;
  }
  
  /**
   * Buyer - Information about the buyer
   */
  export interface Buyer {
    /** First name of buyer */
    first_name: string;
    /** Last name of buyer */
    last_name: string;
    /** Email address of buyer to be used for communication */
    email: string;
    /** Optional phone number of the buyer */
    phone_number?: string;
  }
  
  /**
   * Item - A piece of merchandise that can be purchased
   */
  export interface Item {
    /** Id of a piece of merchandise that can be purchased */
    id: string;
    /** Quantity of the item for fulfillment (minimum 1) */
    quantity: number;
  }
  
  /**
   * Payment provider types
   */
  export type PaymentProviderType = 'frisbii';
  
  /**
   * Supported payment methods
   */
  export type PaymentMethodType = 'card';
  
  /**
   * PaymentProvider - Payment processor configuration
   */
  export interface PaymentProvider {
    /** String value representing payment processor */
    provider: PaymentProviderType;
    /** List of payment methods that the merchant is willing to accept */
    supported_payment_methods: PaymentMethodType[];
  }
  
  /**
   * LineItem - Item in a shopping cart or order
   */
  export interface LineItem {
    /** Id of the line item (different from the item id) */
    id: string;
    /** Item that is represented by the line item */
    item: Item;
    /** Integer representing item base amount before adjustments */
    base_amount: number;
    /** Integer representing any discount applied to the item */
    discount: number;
    /** Integer representing amount after all adjustments */
    subtotal: number;
    /** Integer representing tax amount */
    tax: number;
    /** Integer representing total amount */
    total: number;
  }
  
  /**
   * Total types
   */
  export type TotalType =
    | 'items_base_amount'
    | 'items_discount'
    | 'subtotal'
    | 'discount'
    | 'fulfillment'
    | 'tax'
    | 'fee'
    | 'total';
  
  /**
   * Total - Price breakdown component
   */
  export interface Total {
    /** String value representing the type of total */
    type: TotalType;
    /** The text displayed to the customer for this total */
    display_text: string;
    /** Integer representing total amount in minor units */
    amount: number;
  }
// ...
  export type FulfillmentOption = ShippingFulfillmentOption | DigitalFulfillmentOption;
// ...
  export type Message = InfoMessage | ErrorMessage;
  
  /**
   * Link types
   */
  export type LinkType = 'terms_of_use' | 'privacy_policy' | 'seller_shop_policies';
  
  /**
   * Link - URL link with specific type
   */
  export interface Link {
    /** Type of the link */
    type: LinkType;
    /** Link content specified as a URL */
    url: string;
  }
  
  /**
   * PaymentData - Payment method information
   */
  export interface PaymentData {
    /** Token that represents the payment method */
    token: string;
    /** String value representing the payment processor */
    provider: PaymentProviderType;
    /** Optional billing address associated with the payment method */
    billing_address?: Address;
  }
  
  /**
   * Order - Completed order information
   */
  export interface Order {
    /** Unique id that identifies the order */
    id: string;
    /** Id that identifies the checkout session that created this order */
    checkout_session_id: string;
    /** URL that points to the order */
    permalink_url: string;
  }
  
  // ============================================================================
  // Checkout Session Types
  // ============================================================================
  
  /**
   * Checkout session status
   */
  export type CheckoutSessionStatus =
    | 'not_ready_for_payment'
    | 'ready_for_payment'
    | 'completed'
    | 'canceled'
    | 'in_progress';

  /**
   * CheckoutStatus - Alias for CheckoutSessionStatus (for compatibility)
   */
  export type CheckoutStatus = CheckoutSessionStatus;

  /**
   * CheckoutSessionBase - Base checkout session properties
   */
  export interface CheckoutSessionBase {
    /** Unique id that identifies the checkout session */
    id: string;
    /** Buyer information, if provided */
    buyer?: Buyer;
    /** Payment provider that will be used to complete this transaction */
    payment_provider?: PaymentProvider;
    /** Current status of the checkout session */
    status: CheckoutSessionStatus;
    /** Currency code as per the ISO 4217 standard (lowercase) */
    currency: string;
    /** List of items and computed costs */
    line_items: LineItem[];
    /** Address to ship items to */
    fulfillment_address?: Address;
    /** All available fulfillment options and associated costs */
    fulfillment_options: FulfillmentOption[];
    /** Id of the selected fulfillment option */
    fulfillment_option_id?: string;
    /** List of totals */
    totals: Total[];
    /** List of informational and error messages to be displayed to the customer */
    messages: Message[];
    /** List of links (e.g., TOS/privacy policy/etc.) to be displayed to the customer */
    links: Link[];
  }
  
  /**
   * CheckoutSession - Standard checkout session response
   */
  export interface CheckoutSession extends CheckoutSessionBase {
    /** Order information after completion (only present when status is 'completed') */
    order?: Order;
  }
// ...
  export interface CheckoutSessionCreateRequest {
    /** Optional information about the buyer */
    buyer?: Buyer;
    /** The initial list of items to initiate the checkout session (non-empty) */
    items: Item[];
    /** Optional fulfillment address if present */
    fulfillment_address?: Address;
  }
  
  /**
   * CheckoutSessionUpdateRequest - Request to update a checkout session
   */
  export interface CheckoutSessionUpdateRequest {
    /** Optional information about the buyer */
    buyer?: Buyer;
    /** Optional list of items to update */
    items?: Item[];
    /** Optional fulfillment address */
    fulfillment_address?: Address;
    /** Optional fulfillment option id to select */
    fulfillment_option_id?: string;
  }
  
  /**
   * CheckoutSessionCompleteRequest - Request to complete a checkout session
   */
  export interface CheckoutSessionCompleteRequest {
    /** Optional buyer information */
    buyer?: Buyer;
    /** Payment data required to complete the transaction */
    payment_data: PaymentData;
  }